feat: separa sdk por domínios
